Fix error-handling logic
The possible returns for the http request make this very complex. We
need to handle things like authentication failures (retry) and other
retryable errors e.g. transient network failures, but also handle cases
where we get a valid HTTP response but it isn't a success. If http.Do
returns an error, then response will be nil. If it doesn't then it won't
and we will also have a body we need to close.