Move prometheus listen port to cluster config
Each cluster needs to run a Prometheus metrics listener on a separate
port. We could use a "base port" and increment but that would make it
hard to know which port maps to which cluster over time.
Add Makefile because I'm lazy.
Improve/add comments.
Fix/update comments in the config file.
Bump version.