A couple of tiny fixes
The InfluxDBv2 API does not return an error type so the check for
'err != nil' that was copied from the v1 code was unnecessary and has
been deleted.
The code to calulate the stat buckets takes a cluster struct as an
argument but wasn't using it. Added cluster name to the messages
output from this function.