Refactor back end to separate OneFS-specfic code
Previously, the backend timeseries-database writer abstraction "knew
about the OneFS stats (/1/statistics/current) and performed the
conversion in each implementation. Now the conversion of stats data to
points is handled once in backend.go and the timeseries backend takes a
generic array of "Point"s. This is in preperation to add support for the
summary statistics which use a different endpoint.

Bump version to 0.24 to distinguish although there is no external
functionality change and the v0.24 should behave identically to v0.23.