Force socket reuseaddr/reuseport on prom endpoints
The Golang net/http library functions such as http.ListenAndServe() do
not seem to have any ability to set the reuseaddr and reuseport options.
This is a problem because if the collector is stopped and then restarted
within 2*MSL, then the bind operation will fail with address/port in
use. This PR adds code that uses a net.ListenConfig{} to set these
options.