Fix prometheus summary stats and tidy metricMap
The new protocol summary stats code was broken for Prometheus.
Clean up/simplify metricMap after the points refactor add an entry for
each of the summary stats (only protocol so far), and fix the subsequent
write points breakage due to one of the fields in the summary stats
structure being defined as int64 instead of float64.