Rewrite stat decoder, add many tests
Greatly expand tests.
Add Makefile entry for test.
Switch to using the maps package which means we now require at least
Golang version 1.21.
Fix "isInvalidStat()" which has been broken then entire time (it was
checking fields but the op_name is a tag, not a field).
Bump version in case there are problems with the changes.