Fix regular stats node vs devid
Historically the collector has pulled regular stats from the
/1/statistics/current endpoint with just '?degraded=true&nodes=all'.
Because of this, it only pulled the internal node devid and that was
used for the "nodr" label for the node-specific statistics.
It was pointed out that the endpoint takes a "show_nodes" boolean
argument which, when passed, returns a "node" label.

This commit changes the labels so the "node" label now actually shows
the LNN (logical node number), and a new "devid" label shows the
internal node devid. For clusters where these match (where no nodes have
every been removed and readded), the numbers will match. For clusters
where they do not match, the logical node number is much more useful.