Fix handling of degraded stats
When a node in the cluster is down, stat returns are marked degraded.
The stat engine will still retrieve data but the stat returns are marked
with an error code. Unfortunately, code to handle missing stats and add
them to a list of "bad" stats that should not be collected also caught
the degraded case and so these stats ended up "disappearing".

Added a full list of the expected return types and handle them all
appropriately. Some are permanent errors which result in the stat being
added to the "bad"/"do not collect" bucket. Some are temporary and
result in the stat just being skipped for this collection cycle. The
"degraded" cases will now continue to process the stat.

All stats now have a "degraded" label/tag that is set to true or false
depending on whether we received a degraded result.

Bump the version to "0.30" but mark as still compatible with older
config files.